+++ src/RaZ/Render/Texture.cpp
@@ -20,13 +20,21 @@ Texture::Texture(unsigned int width, unsigned int height, bool isDepthTexture) :
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, width, height, 0, GL_RGB, GL_UNSIGNED_BYTE, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, static_cast<int>(width), static_cast<int>(height), 0, GL_RGB, GL_UNSIGNED_BYTE, nullptr);
     glGenerateMipmap(GL_TEXTURE_2D);
   } else {
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT32F, width, height, 0, GL_DEPTH_COMPONENT, GL_FLOAT, nullptr);
+    glTexImage2D(GL_TEXTURE_2D,
+                 0,
+                 GL_DEPTH_COMPONENT32F,
+                 static_cast<int>(width),
+                 static_cast<int>(height),
+                 0,
+                 GL_DEPTH_COMPONENT,
+                 GL_FLOAT,
+                 nullptr);
   }
 
   unbind();
@@ -63,7 +71,7 @@ void Texture::load(const std::string& fileName) {
 
   glTexImage2D(GL_TEXTURE_2D,
                0,
-               img.getColorspace(),
+               static_cast<int>(img.getColorspace()),
                static_cast<int>(img.getWidth()),
                static_cast<int>(img.getHeight()),
                0,
